requirements.txt
django
djangorestframework
djangorestframework-simplejwt
channels
channels-redis
asgiref
sqlparse
redis
PyJWT
msgpack
drf_spectacular
daphne
gunicorn
